+++ src/spring-cloud/HISTORY.md
@@ -1,5 +1,10 @@
 Release History
 ===============
+2.1.1
+-----
+* Remove preview parameter '--enable-java-agent' from 'az spring-cloud update'.
+* Fix warning message of '--disable-distributed-tracing'.
+
 2.1.0
 -----
 * Support Java In-Process Agent.